% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seabird.R
\name{scale_color_seabird}
\alias{scale_color_seabird}
\title{Color scale for Seabird colors}
\usage{
scale_color_seabird(
  palette = "bluefootedbooby",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in seabird_palettes}

\item{discrete}{Boolean if color aesthetic is discrete}

\item{reverse}{Boolean indicating whether palette should be reversed}

\item{...}{Additional arguments used to discrete_scale() or scale_fill_gradientn()
to automatically interpolate between colours.}
}
\description{
Color scale for Seabird colors
}
