% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seabird.R
\name{seabird_pal}
\alias{seabird_pal}
\title{Function to interpolate a color palette}
\usage{
seabird_pal(palette = "bluefootedbooby", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in seabird_palettes}

\item{reverse}{Boolean true if palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Function to interpolate a color palette
}
