% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seabird.R
\docType{data}
\name{seabird_palettes}
\alias{seabird_palettes}
\title{Color palettes}
\format{
An object of class \code{list} of length 8.
}
\usage{
seabird_palettes
}
\description{
Color palettes
}
\keyword{datasets}
